separate class into package
