+++ single-table-inheritance/src/main/java/com/iluwatar/SingleTableInheritance/CarMapper.java
@@ -1,9 +1,8 @@
 package com.iluwatar.SingleTableInheritance;
 
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.EntityTransaction;
-import javax.persistence.Persistence;
+import com.iluwatar.SingleTableInheritance.ClassObject.Car;
+import com.iluwatar.SingleTableInheritance.ClassObject.Vehicle;
+
 import java.util.List;
 
 public class CarMapper extends AbstractVehicleMapper<Car>{
@@ -15,6 +14,10 @@ public class CarMapper extends AbstractVehicleMapper<Car>{
 
     public Car Load(List<Vehicle> rows){
         return (Car) super.Load(rows);
+    }
+
+    public Car Save(Vehicle v, boolean update){
+        return (Car) super.Save(v, update);
     }
 
     public Car Update(Vehicle v){
