+++ single-table-inheritance/src/test/java/com/iluwatar/SingleTableInheritance/SingleTableInheritanceTest.java
@@ -1,5 +1,6 @@
 package com.iluwatar.SingleTableInheritance;
 
+import com.iluwatar.SingleTableInheritance.ClassObject.*;
 import org.junit.Assert;
 import org.junit.Test;
 import javax.persistence.*;
